+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Fix SA Mart threads posting with the wrong secondary tag.
 * Prevent Twitter embeds from hijacking posts web view.
+* Restore thread tags to The Film Dump.
 
 ## 3.17